Move nonce related calculations into protocol rules
This addesses one part of #946

All nonce updates are now managed in UPDN, which has an additional case when the slot is the first slot of the epoch, updating the epoch nonce. All mention of nonces is now removed from the ledger part of the rules.

I think this is actually a nicer formulation after all anyway.